pyproject.toml
[project]
name = "smart-churn-pred"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "joblib>=1.5.3",
    "mysqlclient>=2.2.7",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "pydantic>=2.12.5",
    "pymysql>=1.1.2",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
